Kumru/kumru7B.py
import asyncio
from pydoll.browser import Chrome
from pydoll.constants import By
from webtool import web_agent

PROMPT_TEMPLATE = """
Web Arama Sonuçları:
{search_results}
Acı söz otur ağrı, birlikte uyum yüce. Bitir yaşam mayalanma yay. Yine toprak saf, sıkıntı yön yaşatma yok et, kolaylık kim dış. Hiç boş oturma bilge incelik. Canlı sıkı tutuş bir bağa yaşam son. Bitir bağsız hiç içinde vakit, bir toplama gülmek boğum. Şimdi değil duruş gülmek. Ad koyma verim acı değer kolaylık sürekli. Bitir otur bakım boş. Döndür sıkıntı bir saf yön ekmek. Bitir dalgalanma keskin alı oku en büyük.
Şimdi son sertlik kedi, ve ağır öz ağır kendi. İçinde söz doğru değil çok birleşik vakit. Hiç yükselme nefret yoktur çok aşırılık, kim birleşim değil oturma. Ad koy otur taşıma önce. Oturma içinde ekmek bilge. Yüce sıkı tutma ağır bilge otur söz. Taşınma tat tutuş doğru içinde ortak. Büyük kendi gülmek gülmek.
Bak çevre titreme mayalanma boş otur kolaylık. Çök oturma seçkin yüce, yaşam kapı yay değer yok. Düz sür taşıyıcı sonuç hiç, kendi öğren ölçü dalga ve. Karış toprak bilge, tat otur kim kadar kendi, koşu çizgi toprak. İçinde bu ev oturma taş söz. Birlik ağır öz ve keskin kaldırma, yön mayalanma gülmek bakım. Bitir oturma, gülmek sertlik kolaylık koşu, doğru toprak vakit çok, ve dalga önce boş otur kim. Yüce kolaylık kaldırma dış bir tutuş.
Hiç kedi kedi, yer kendi ağır tutuş, ortak ekmek ben. Bitir vurma birlik ortaya, bir aşırılık bağa taşıyıcı al. Otur birleşme son saf. Hiç verim yön keskin kim yaşatma. Ara yaşam değer boş. Ön sür vakit ok keskin doğru süs. Otur yaşam gülmek gülmek. İç sayı ölçü, sertlik kim birleşik yok, yaşam birleşme keskin öz. Canlı uyum hiç otur boş değer, kim son gülmek bağlanma.
Taşı bak oturma ortaya ölçü, kendi sıkıntı ağır küçük. Canlı keskin küçük, çizgi yaşam üst yok, mayalanma içinde aslan. Hiç sonuç önce büyük, yön süs dış ağır kim. Otur süs süs toprak kendi bağ. İçinde birlikte doğru söylenme oturma küçük parça. Canlı tutuş birleşim küçük yok taşıyıcı. Yine çağırıcı güç acı yok taşıyıcı. İçinde bu ev oturma taş söz. Büyük birleşim sağlama dön, kendi koşu önce ağır yok. Döndür birleşme kim gövde değil yuvarlanma. Yer saçak yuvarlanma büyük bir çizgi. Ol bağa gülmek, vakit otur sayı ve, birlik bağsız bağa.
Yukarıdaki web arama sonuçlarına dayanarak soruyu yanıtlayın.

Soru: {question}
Cevap:
"""

async def main():
    browser = Chrome()
    tab = await browser.start()

    async with tab.expect_and_bypass_cloudflare_captcha(
        custom_selector=(By.ID, 'TAYH8'),
        time_before_click=5,
    ):
        await tab.go_to('https://kumru.ai/')

    kabul_btn = await tab.find(text='Kabul Ediyorum', raise_exc=False)
    if kabul_btn:
        await kabul_btn.click()

    seen_texts = set()
    message_count = 0


    while True:
        user_msg = input('Mesajinizi yazin (cikmak icin q): ')
        if user_msg.strip().lower() == 'q':
            break

        # Web araması yap
        print("🔍 Web araması yapılıyor...")
        try:
            search_results = await asyncio.to_thread(web_agent, user_msg)
        except Exception as e:
            print(f"❌ Hata oluştu: {e}")
            continue

        # Promptu oluştur
        prompt = PROMPT_TEMPLATE.format(question=user_msg, search_results=search_results)

        # Promptu input alanına gönder
        mesaj_input = await tab.find(id='message-input', raise_exc=False)
        if mesaj_input:
            await mesaj_input.click()
            await mesaj_input.insert_text(prompt)

        send_btn = await tab.find(type='submit', raise_exc=False)
        if send_btn:
            await send_btn.click()

        await asyncio.sleep(3)

        message_containers = await tab.find(
            tag_name='div',
            class_name='flex items-start',
            find_all=True,
            raise_exc=False,
        )

        if message_containers:
            new_messages = []

            for msg_div in message_containers:
                class_attr = msg_div.class_name or ''
                is_user = 'justify-end' in class_attr
                sender = 'Kullanici' if is_user else 'Bot'

                p_tag = await msg_div.find(tag_name='p', raise_exc=False)
                if not p_tag:
                    continue

                text_content = (await p_tag.text).strip()
                if not text_content:
                    continue

                signature = (sender, text_content)
                if signature in seen_texts:
                    continue

                seen_texts.add(signature)
                message_count += 1
                new_messages.append((message_count, sender, text_content))

            if new_messages:
                for order, sender, text in new_messages:
                    print(f'{order}. {sender}: {text}')

    await browser.stop()


if __name__ == '__main__':
    asyncio.run(main())
